+++ packages/huma-shared/src/v2/utils/pool.ts
@@ -78,12 +78,17 @@ export type PoolInfoV2 = {
   lenderApprovalProvider?: LenderApprovalProvider
   industry: IndustryType
   KYC?: {
-    provider: 'Securitize'
-    signInRequired: KYCCopy
-    verifyIdentity: KYCCopy
-    emailSignatureLink: KYCCopy
-    resendSignatureLink: KYCCopy
-    docUnderReview: KYCCopy
+    Securitize?: {
+      signInRequired: KYCCopy
+      verifyIdentity: KYCCopy
+      emailSignatureLink: KYCCopy
+      resendSignatureLink: KYCCopy
+      docUnderReview: KYCCopy
+    }
+    Persona?: {
+      signInRequired: KYCCopy
+      verifyIdentity: KYCCopy
+    }
   }
   supplyLink?: string
   poolUnderlyingToken: {
