+++ packages/huma-widget/package.json
@@ -63,6 +63,7 @@
     "history": "^5.3.0",
     "jotai": "^1.13.1",
     "moment": "^2.29.4",
+    "persona": "^5.1.0",
     "qrcode": "^1.5.1",
     "query-string": "^7.1.1",
     "react": "^18.2.0",
@@ -137,6 +138,7 @@
     "@typechain/ethers-v5": "^10.1.0",
     "@types/jest": "^27.5.2",
     "@types/node": "^17.0.45",
+    "@types/persona": "^0.0.32",
     "@types/qrcode": "^1.5.0",
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
